# spiffworkflow-backend/src/spiffworkflow_backend/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: spiffworkflow-backend
  license:
    name: LGPL
servers:
  - url: http://localhost:5000/v1.0
# this is handled in flask now
security: []
paths:
...
  /task-assign/{modified_process_model_identifier}/{process_instance_id}/{task_guid}:
    parameters:
      - name: modified_process_model_identifier
        in: path
        required: true
        description: The modified id of an existing process model
        schema:
          type: string
      - name: process_instance_id
        in: path
        required: true
        description: The unique id of an existing process instance.
        schema:
          type: integer
      - name: task_guid
        in: path
        required: true
        description: The unique id of the task.
        schema:
          type: string
    post:
      operationId: spiffworkflow_backend.routes.tasks_controller.task_assign
      summary: Assign a given task to a list of additional users
      tags:
        - Process Instances
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "ok: true"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
...
  /tasks/{process_instance_id}:
    parameters:
      - name: process_instance_id
        in: path
        required: true
        description: The unique id of an existing process instance.
        schema:
          type: integer
      - name: execute_tasks
        in: query
        required: false
        description: Execute ready tasks on the process instance.
        schema:
          type: boolean
    get:
      tags:
        - Tasks
      operationId: spiffworkflow_backend.routes.tasks_controller.interstitial
      summary: An SSE (Server Sent Events) endpoint that returns what tasks are currently active (running, waiting, or the final END event)
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /tasks/{process_instance_id}/{task_guid}:
    parameters:
      - name: task_guid
        in: path
        required: true
        description: The unique id of an existing process group.
        schema:
          type: string
      - name: process_instance_id
        in: path
        required: true
        description: The unique id of an existing process instance.
        schema:
          type: integer
      - name: with_form_data
        in: query
        required: false
        description: Include task data for forms
        schema:
          type: boolean
      - name: execution_mode
        in: query
        required: false
        description: Either run in "synchronous" or "asynchronous" mode.
        schema:
          type: string
          enum:
            - synchronous
            - asynchronous
    get:
      tags:
        - Tasks
      operationId: spiffworkflow_backend.routes.tasks_controller.task_show
      summary: Gets one task that a user wants to complete
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      tags:
        - Tasks
      operationId: spiffworkflow_backend.routes.tasks_controller.task_submit
      summary: Update the form data for a tasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "202":
          description: "ok: true"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
...
  /public/tasks/{process_instance_id}/{task_guid}:
    parameters:
      - name: task_guid
        in: path
        required: true
        description: The unique id of an existing process group.
        schema:
          type: string
      - name: process_instance_id
        in: path
        required: true
        description: The unique id of an existing process instance.
        schema:
          type: integer
      - name: execution_mode
        in: query
        required: false
        description: Either run in "synchronous" or "asynchronous" mode.
        schema:
          type: string
          enum:
            - synchronous
            - asynchronous
    get:
      tags:
        - Tasks
      operationId: spiffworkflow_backend.routes.public_controller.form_show
      summary: Gets one task that a user wants to complete
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      tags:
        - Tasks
      operationId: spiffworkflow_backend.routes.public_controller.form_submit
      summary: Update the form data for a tasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "202":
          description: "ok: true"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
...
  /task-filters:
    post:
      summary: Retrieve filtered tasks based on pagination and other criteria
      operationId: spiffworkflow_backend.routes.ff_tasks_controller.filter_tasks
      parameters:
        - name: firstResult
          in: query
          required: true
          description: The starting index of the results to return
          schema:
            type: integer
        - name: maxResults
          in: query
          required: true
          description: The maximum number of results to return
          schema:
            type: integer
      responses:
        "200":
          description: Successfully retrieved filtered tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _links:
                      type: object
                      properties:
                        self:
                          type: object
                          properties:
                            href:
                              type: string
                    _embedded:
                      type: object
#                      properties:
#                        assignee:
#                          type: array
#                          items:
#                            $ref: '#/components/schemas/FfUser'
#                        processDefinition:
#                          type: array
#                          items:
#                            $ref: '#/components/schemas/ProcessDefinition'
#                        task:
#                          type: array
#                          items:
#                            $ref: '#/components/schemas/FfTask'
                    count:
                      type: integer
  /task/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        description: "The unique id of an existing process group."
        schema:
          type: string
    get:
      tags:
        - Tasks
      operationId: spiffworkflow_backend.routes.ff_tasks_controller.get_task_by_id
      summary: "Gets one task that a user wants to complete"
      responses:
        "200":
          description: "One task"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /task/{task_id}/claim:
    post:
      summary: "Claim task"
      operationId: spiffworkflow_backend.routes.ff_tasks_controller.claim_task
      parameters:
        - name: task_id
          in: path
          required: true
          description: "The unique id of an existing process group."
          schema:
            type: string
      responses:
        "200":
          description: "Successfully claimed task"
          content:
            application/json:
              schema:
                type: object
  /task/{task_id}/unclaim:
    post:
      summary: "Unclaim task"
      operationId: spiffworkflow_backend.routes.ff_tasks_controller.unclaim_task
      parameters:
        - name: task_id
          in: path
          required: true
          description: "The unique id of an existing process group."
          schema:
            type: string
      responses:
        "200":
          description: "Successfully unclaimed task"
          content:
            application/json:
              schema:
                type: object
  /task/{task_id}/submit-form:
    post:
      summary: "Submit form"
      operationId: spiffworkflow_backend.routes.ff_tasks_controller.submit_task
      parameters:
        - name: task_id
          in: path
          required: true
          description: "The unique id of an existing process group."
          schema:
            type: string
      responses:
        "200":
          description: "Successfully submitted task"
          content:
            application/json:
              schema:
                type: object
components:
  securitySchemes:
    oAuth2AuthCode:
      type: oauth2
      description: authenticate with openid server
      flows:
        authorizationCode:
          authorizationUrl: /v1.0/login_api
          tokenUrl: /v1.0/login_api_return
          scopes:
            read_email: read email
      x-tokenInfoFunc: spiffworkflow_backend.routes.authentication_controller.get_scope
  schemas:
    OkTrue:
      properties:
        ok:
          type: boolean
          example: true
...
    User:
      properties:
        uid:
          type: string
        email_address:
          type: string
        display_name:
          type: string
        affiliation:
          type: string
        eppn:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        title:
          type: string
...
    Task:
      properties:
        id:
          readOnly: true
          type: string
        name:
          type: string
        title:
          type: string
        type:
          type: string
        state:
          type: string
        documentation:
          type: string
        data:
          type: object
        multi_instance_type:
          type: string
          enum: ["none", "looping", "parallel", "sequential"]
        multi_instance_count:
          type: number
        multi_instance_index:
          type: number
        process_name:
          type: string
        properties:
          type: object
      example:
        id: study_identification
        name: Study Identification
        title: IRB Review
        documentation: "# Heading 1\n\nMarkdown documentation text goes here"
        type: form
        state: ready
        form:
          "key": "irb_review_form"
          "fields":
            - "id": "irb_review_type"
              "type": "enum"
              "label": "Select IRB Review Type"
              "options":
                - id: "emergency_use"
                  name: "Emergency Use"
                - id: "humanitarian_device"
                  name: "Humanitarian Device"
                - id: "non_human"
                  name: "Non-Human"
                - id: "non_uva_agent"
                  name: "Non-UVA Agent"
                - id: "exempt"
                  name: "Exempt"
                - id: "non_engaged"
                  name: "Non-Engaged"
                - id: "expedited"
                  name: "Expedited"
                - id: "full_board"
                  name: "Full Board"
              "default_value": "Full Board"
              "validation":
                - name: "required"
                  config: "true"
              "properties":
                - id: "description"
                  value: "Description text goes here"
                - id: "help"
                  value: "# Heading 1\n\nMarkdown help text goes here"
                - id: "required_expression"
                  value: "model.my_boolean_field_id && model.my_enum_field_value !== 'something'"
                - id: "hide_expression"
                  value: "model.my_enum_field_value === 'something'"
...
    AwesomeUnspecifiedPayload:
      # we know that task_submit submits no body at all, and None is not an object, as this so helpfully tells us
      # type: "object"
      additionalProperties: {}
...